Allow to disable "Add OpenJDK-R Java PPA" task
When using the role behind a corporate firewall, or when intending to
install Java only from the packages available in the distribution, we
don't want to add the OpenJDK PPA to the repositories list.

Add a new variable, java_openjdk_use_ppa, which will prevent this task
from running when set to false. The default value is true, meaning the
default behaviour is unchanged.

Change-Id: I087c95f0742925a7adf998c9a563489526dd6c56